refactor(calendar): split worldcalendarconfigpanel into smaller modules
Split the panel into focused calendar feature modules: a thin query wrapper, editable form, list
editor, read-only summary, field primitives if needed, and validation/error helpers.

fix #120
